requirements.txt
click
gitpython
pygithub
google-generativeai
flake8
pytest
coverage
flask
gunicorn
